Drive fault conditions with unique timers
Refactor to remove the dependency on the global millisecond timer.

Previously, the fault condition mechanism relied on a global millisecond
timer to provide a common time domain and was vulnerable to u32 wrapping problems.
It was a common point of failure, if something messed up the data watchpoint trigger (DWT)
configuration then faults could go undetected.

This commit removes the millisecond timer and it's associated time domain arithmetic,
and instead provides each instantiation of a FaultCondition a unique count down
peripheral timer that implements the One-Pulse trait.

The One-Pulse timer period is configured at board/peripheral construction time, as the
timeout value is constant.