Add default panic implementation
Add a default panic implementation that performs the following (in order):

- disable safety/control related GPIO pins
- enable red LED and brake lights as a fault indicator
- output the PanicInfo to Serial3
- call `intrinsics::abort`

**NOTE**
The watchdog is enable by default, so you might not even notice the fault.
It might be worth exposing the `hard_fault_indicator()` function so we
can call it during initialization if the `ResetConditions` indicate
a watchdog reset. That would persist the indication across panicing/faults.
We could also disable the watchdog (and other interrupts) completely to stay in the abort.